ssh_auth_podman_push_lib
Module encrypt_decrypt_with_ssh_key_mod

The mod.rs generic name is difficult to find and maintain.

Here I will have only code that uses other modules with meaningful names. Don’t change this code, so it can be updated regularly with cargo auto update_automation_tasks_rs If you want to customize it, copy the code into main.rs and modify it there.

Modules

docker_io_api_token_mod

decrypt docker.io api token from file or ask the user to input the access token, encrypt it and save into file

encrypt_decrypt_mod

Generic functions to encrypt and decrypt secrets using the ssh private key.